Reduce the recipe by leveraging new recipes
After simmering for a few days we have now reduced the mixture.
Minding the metaphor we now have a bunch of new recipes such as
`silverstripe/recipe-blog` that we can use instead of inlining all
of it's requirements directly in this recipe's composer.json

Along with this some previously removed (via misunderstanding of the
purpose of this new version of this installer recipe) functionality
being added back in by nature of still being necessary to keep in line
with the previous major version.