Allow developers to better use _ss_environment.php
If a developer has several projects set up, it is likely they share environment settings between them. The configuration here however disallows the ease-of-use feature of _ss_environment.php to select a database name depending on the (grand)*parent folder's name.
So we should also check for this before dictating a fallback database name.